Ok I think I actually figured out what was happening for reals
So that line is to register an event listener for battery changes. I guess I didn't realize it at the time because I was using an example, and not really understanding these lines...

My guess as to how the app crash/hang occurs must first meet the following conditions:
- The event has to be registered multiple times. This is done either clicking the refresh button a lot, and/or multiple app resumes (due to lock/unlock, or navigating back to the app and it hasn't been killed off in background yet, etc.) Then, the battery needs to have changed status, often enough to trigger the report. I'm assuming this is why the hang/crash happens over time, because the event doesn't get listened to while the app is in the background. With each resume, the reports (battery changes) that happened were queued up(?) and are now triggering the event listener, which has been registered with my method multiple times, AND to top it off, the event listener also registers itself again too! So then it calls these new handlers until a race condition probably puts a halt to this, where it isn't able to register itself in time before it considered itself done calling all the registered methods (it's run asynchronously). Repeat this cycle multiple times, where the listener is registered hundreds of times, and then on resume... it hangs and goes boom!

My last recorded count of a crash recorded 263 entries to the log file... epic.

Idk if I explained the issue well enough there. tl;dr event listener was registering itself, async conditions made this hard to find as the culprit.